+++ src/main.tsx
@@ -1,12 +1,16 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
+import ThemeColorsPage from "./ThemeColorsPage";
 import { MedixProvider } from "../lib";
+
+const normalizedPath = window.location.pathname.replace(/\/+$/, "") || "/";
+const previewPage = normalizedPath === "/theme-colors" ? <ThemeColorsPage /> : <App />;
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <MedixProvider defaultColorMode="light">
-      <App />
+      {previewPage}
     </MedixProvider>
   </React.StrictMode>
 );